refactor util func
